move context definition to settings component folder
